feat: implement Redux Toolkit for home page
